refactor: backend

+++ backend/apps/users/routers.py
@@ -9,7 +9,6 @@ from sqlalchemy.orm import selectinload
 import settings
 from dependencies import get_session
 from apps.core.models import Image
-from apps.core.utils import handle_file_upload
 from apps.auth.utils import get_hashed_password, verify_password
 from apps.auth.dependencies import get_current_active_user
 from apps.events.models import Event, Sign
@@ -21,42 +20,38 @@ from . import crud
 router = APIRouter()
 
 
-@router.get("/users/me/", tags=['me'], response_model=UserRetrieve)
-async def get_user_me(current_user: UserRetrieve = Security(get_current_active_user, scopes=['me'])):
+@router.get("/users/me/", tags=['profile'], response_model=UserRetrieve)
+async def get_user_me(current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile'])):
     return current_user
 
 
-@router.put('/users/me/', tags=['me'])
+@router.put('/users/me/', tags=['profile'], response_model=UserRetrieve)
 async def update_user(user: UserUpdate,
-                      current_user: UserRetrieve = Security(
-                          get_current_active_user, scopes=['me']),
+                      current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile']),
                       session: AsyncSession = Depends(get_session)):
-    current_user = await crud.update_user(user, current_user, session)
-    return JSONResponse({}, status_code=status.HTTP_200_OK)
+    return await crud.update_user(user, current_user, session)
 
 
-@router.put('/users/me/photo/', tags=['me'])
+@router.put('/users/me/photo/', tags=['profile'], response_model=ImageBase)
 async def update_photo(image_id: int,
-                       current_user: UserRetrieve = Security(get_current_active_user, scopes=['me']),
+                       current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile']),
                        session: AsyncSession = Depends(get_session)):
-    try:
-        image_obj = (await session.execute(select(Image)
-                                           .where(Image.id == image_id))).scalar()
-        image_obj.object_type = 'User'
-        image_obj.object_id = current_user.id
-    finally:
+    image_obj = (await session.execute(select(Image)
+                                       .where(Image.id == image_id))).scalar()
+    image_obj.object_type = 'User'
+    image_obj.object_id = current_user.id
+    if current_user.image and current_user.image.id != image_id:
         file_path = os.path.join(settings.BASEDIR, current_user.image.url[1:])
         if os.path.exists(file_path):
             os.remove(file_path)
         await session.execute(delete(Image).where(Image.id == current_user.image.id))
     await session.commit()
-    return JSONResponse({}, status_code=status.HTTP_200_OK)
+    return image_obj
 
 
-@router.delete('/users/me/photo/', tags=['me'])
-async def delete_photo(current_user: UserRetrieve = Security(
-        get_current_active_user, scopes=['me']),
-        session: AsyncSession = Depends(get_session)):
+@router.delete('/users/me/photo/', tags=['profile'])
+async def delete_photo(current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile']),
+                       session: AsyncSession = Depends(get_session)):
     file_path = os.path.join(settings.BASEDIR, current_user.image.url[1:])
     if os.path.exists(file_path):
         os.remove(file_path)
@@ -65,10 +60,9 @@ async def delete_photo(current_user: UserRetrieve = Security(
     return JSONResponse({}, status_code=status.HTTP_202_ACCEPTED)
 
 
-@router.post('/users/me/change_password/', tags=['me'])
+@router.post('/users/me/change_password/', tags=['profile'])
 async def change_password(password_form: UserPasswordChange,
-                          current_user: UserRetrieve = Security(
-                              get_current_active_user, scopes=['me']),
+                          current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile']),
                           session: AsyncSession = Depends(get_session)):
     if verify_password(password_form.old_password, current_user.password):
         current_user.password = get_hashed_password(password_form.new_password)
@@ -79,15 +73,15 @@ async def change_password(password_form: UserPasswordChange,
                             detail="Incorrect old password")
 
 
-@router.get('/users/me/notifications/', tags=['me'])
-async def get_notifications(current_user: UserRetrieve = Security(get_current_active_user, scopes=['me']),
+@router.get('/users/me/notifications/', tags=['profile'])
+async def get_notifications(current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile']),
                             session: AsyncSession = Depends(get_session)):
     return (await session.scalars(current_user.notifications)).all()
 
 
 @router.get('/users/{username}/', tags=['users'])
 async def get_user(username: str,
-                   current_user: UserRetrieve = Security(get_current_active_user, scopes=['me']),
+                   current_user: UserRetrieve = Security(get_current_active_user, scopes=['users']),
                    session: AsyncSession = Depends(get_session)):
     subq_events = (select(func.count('*')).select_from(Event)
                    .where(Event.created_by == current_user.id)
@@ -108,7 +102,7 @@ async def get_user(username: str,
 
 @router.get('/users/{username}/followers/', tags=['users'], response_model=list[UserRetrieve])
 async def get_followers(username: str,
-                        current_user: UserRetrieve = Security(get_current_active_user, scopes=['me', 'users']),
+                        current_user: UserRetrieve = Security(get_current_active_user, scopes=['users']),
                         session: AsyncSession = Depends(get_session)):
     user_obj = await crud.get_user_or_404(username=username, session=session)
     return (await session.scalars(user_obj.followers.statement)).all()
@@ -116,7 +110,7 @@ async def get_followers(username: str,
 
 @router.get('/users/{username}/following/', tags=['users'], response_model=list[UserRetrieve])
 async def get_following(username: str,
-                        current_user: UserRetrieve = Security(get_current_active_user, scopes=['me', 'users']),
+                        current_user: UserRetrieve = Security(get_current_active_user, scopes=['users']),
                         session: AsyncSession = Depends(get_session)):
     user_obj = await crud.get_user_or_404(username=username, session=session)
     return (await session.scalars(user_obj.following.statement)).all()
@@ -124,7 +118,7 @@ async def get_following(username: str,
 
 @router.get('/users/{username}/is_follow/', tags=['users'])
 async def is_following(username: str,
-                       current_user: UserRetrieve = Security(get_current_active_user, scopes=['me', 'users']),
+                       current_user: UserRetrieve = Security(get_current_active_user, scopes=['users']),
                        session: AsyncSession = Depends(get_session)):
     user = await crud.get_user_or_404(username=username, session=session)
     if await crud.check_follow(user, current_user, session):
@@ -135,7 +129,7 @@ async def is_following(username: str,
 
 @router.post('/users/{username}/follow/', tags=['users'])
 async def follow(username: str,
-                 current_user: UserRetrieve = Security(get_current_active_user, scopes=['me', 'users']),
+                 current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile', 'users']),
                  session: AsyncSession = Depends(get_session)):
     user = await crud.get_user_or_404(username=username, session=session)
     if await crud.check_follow(user, current_user, session):
@@ -150,7 +144,7 @@ async def follow(username: str,
 
 @router.post('/users/{username}/unfollow/', tags=['users'])
 async def unfollow(username: str,
-                   current_user: UserRetrieve = Security(get_current_active_user, scopes=['me', 'users']),
+                   current_user: UserRetrieve = Security(get_current_active_user, scopes=['profile', 'users']),
                    session: AsyncSession = Depends(get_session)):
     user = await crud.get_user_or_404(username=username, session=session)
     subscribe_obj = (await session.execute(select(Subscription)
@@ -168,7 +162,7 @@ async def unfollow(username: str,
 @router.get('/users/{username}/events/', tags=['users'])
 async def get_user_events(role: str,
                           username: str,
-                          current_user: UserRetrieve = Security(get_current_active_user, scopes=['events', 'users']),
+                          current_user: UserRetrieve = Security(get_current_active_user, scopes=['users']),
                           session: AsyncSession = Depends(get_session)):
     user = await crud.get_user_or_404(username=username, session=session)
     return (await session.scalars(user.signs.statement
